Change welcome page
upload many .css .js files and imgs, some are not neccessary but kept just in case. The new welcome page is index.html now, the old welcome.html hasn't been changed. A large part of this templete is banned and commented by me. If you think it's useful, uncomment that.